# manual/static/openapi/admin-api-openapi.json
{
  "openapi": "3.0.2",
...
  "info": {
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "contact": {
      "name": "MAIF open source team",
      "email": "oss@maif.fr"
    },
    "description": "Admin API of Daikoku",
    "title": "Daikoku Admin API",
    "version": "16.3.0"
  },
...
  "components": {
    "schemas": {
...
      "done": {
        "description": "task is done",
        "properties": {
          "done": {
            "type": "boolean"
          }
        }
      },
      "error": {
        "description": "error response",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "patch": {
        "description": "A set of changes described in JSON Patch format: http://jsonpatch.com/ (RFC 6902)",
        "type": "array",
        "items": {
          "type": "object",
          "required": [
            "op",
            "path"
          ],
          "properties": {
            "op": {
              "type": "string",
              "enum": [
                "add",
                "replace",
                "remove",
                "copy",
                "test"
              ]
            },
            "path": {
              "type": "string"
            },
            "value": {}
          }
        }
      },
...
      "User": {
        "description": "User",
        "properties": {
          "_id": {
            "type": "string"
          },
          "_deleted": {
            "type": "boolean"
          },
          "tenants": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "origins": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string",
              "enum": [
                "local",
                "otoroshi",
                "ldap",
                "oauth2"
              ]
            }
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "picture": {
            "type": "string",
            "nullable": true
          },
          "pictureFromProvider": {
            "type": "boolean",
            "nullable": true
          },
          "personalToken": {
            "type": "string",
            "nullable": true
          },
          "isDaikokuAdmin": {
            "type": "boolean",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true,
            "description": "optionnal value but required in case of tenant has a Local AuthProvider"
          },
          "hardwareKeyRegistrations": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object"
            }
          },
          "lastTenant": {
            "type": "string",
            "nullable": true
          },
          "metadata": {
            "type": "object",
            "nullable": true
          },
          "defaultLanguage": {
            "type": "string",
            "nullable": true
          },
          "starredApis": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "twoFactorAuthentication": {
            "$ref": "#/components/schemas/TwoFactorAuthentication",
            "nullable": true
          },
          "invitation": {
            "$ref": "#/components/schemas/UserInvitation",
            "nullable": true
          }
        },
        "required": [
          "_id",
          "name",
          "email"
        ]
      },
...
      "TwoFactorAuthentication": {
        "description": "TwoFactorAuthentication",
        "properties": {
          "enabled": {
            "type": "boolean"
          },
          "secret": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "backupCodes": {
            "type": "string"
          }
        },
        "required": [
          "enabled",
          "secret",
          "token",
          "backupCodes"
        ]
      },
...
    }
  },
  "paths": {
    "/admin-api/tenants/{id}": {
      "delete": {
        "summary": "delete a tenant",
        "operationId": "tenants.delete",
        "responses": {
          "200": {
            "description": "entity deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/done"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The ID of the tenant to delete",
            "required": true
          }
        ],
        "tags": [
          "tenant"
        ]
      },
      "patch": {
        "summary": "update a tenant with JSON patch or by  merging JSON object",
        "operationId": "tenants.patch",
        "requestBody": {
          "description": "the patch to update the tenant or a JSON object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/patch"
                  },
                  {
                    "$ref": "#/components/schemas/Tenant"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Tenant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The tenant ID to update",
            "required": true
          }
        ],
        "tags": [
          "tenant"
        ]
      },
      "put": {
        "summary": "update a tenant",
        "operationId": "tenants.update",
        "requestBody": {
          "description": "the ID of the tenant to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Tenant"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Tenant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The ID of the tenant to update",
            "required": true
          }
        ],
        "tags": [
          "tenant"
        ]
      },
      "get": {
        "summary": "read a tenant",
        "operationId": "tenants.findById",
        "responses": {
          "200": {
            "description": "found entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Tenant"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "description": "The ID of the tenant to retrieve",
            "required": true
          }
        ],
        "tags": [
          "tenant"
        ]
      }
    },
...
    "/admin-api/users/{id}": {
      "delete": {
        "summary": "delete a user",
        "operationId": "users.delete",
        "responses": {
          "200": {
            "description": "entity deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/done"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "patch": {
        "summary": "update a user with JSON patch or by  merging JSON object",
        "operationId": "users.patch",
        "requestBody": {
          "description": "the patch to update the user or a JSON object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/patch"
                  },
                  {
                    "$ref": "#/components/schemas/User"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "put": {
        "summary": "update a user",
        "operationId": "users.update",
        "requestBody": {
          "description": "the user to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      },
      "get": {
        "summary": "read a user",
        "operationId": "users.findById",
        "responses": {
          "200": {
            "description": "found entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "user"
        ]
      }
    },
...
    "/admin-api/teams/{id}": {
      "delete": {
        "summary": "delete a team",
        "operationId": "teams.delete",
        "responses": {
          "200": {
            "description": "entity deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/done"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "team"
        ]
      },
      "patch": {
        "summary": "update a team with JSON patch or by  merging JSON object",
        "operationId": "teams.patch",
        "requestBody": {
          "description": "the patch to update the team or a JSON object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/patch"
                  },
                  {
                    "$ref": "#/components/schemas/Team"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "team"
        ]
      },
      "put": {
        "summary": "update a team",
        "operationId": "teams.update",
        "requestBody": {
          "description": "the team to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Team"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "team"
        ]
      },
      "get": {
        "summary": "read a team",
        "operationId": "teams.findById",
        "responses": {
          "200": {
            "description": "found entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Team"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "team"
        ]
      }
    },
...
    "/admin-api/apis/{id}": {
      "delete": {
        "summary": "delete a api",
        "operationId": "apis.delete",
        "responses": {
          "200": {
            "description": "entity deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/done"
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "api"
        ]
      },
      "patch": {
        "summary": "update a api with JSON patch or by  merging JSON object",
        "operationId": "apis.patch",
        "requestBody": {
          "description": "the patch to update the api or a JSON object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/patch"
                  },
                  {
                    "$ref": "#/components/schemas/Api"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Api"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "api"
        ]
      },
      "put": {
        "summary": "update a api",
        "operationId": "apis.update",
        "requestBody": {
          "description": "the api to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Api"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "updated entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Api"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "api"
        ]
      },
      "get": {
        "summary": "read a api",
        "operationId": "apis.findById",
        "responses": {
          "200": {
            "description": "found entity",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "items": {
                    "$ref": "#/components/schemas/Api"
                  }
                }
              }
            }
          },
          "401": {
            "description": "unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          },
          "404": {
            "description": "entity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "tags": [
          "api"
        ]
      }
    },
...
  }
}